Further transitioning to new Podfile structure.

+++ Kiosk/Bid Fulfillment/PlaceBidViewController.swift
@@ -147,7 +147,7 @@ private extension PlaceBidViewController {
     }
 
     func deleteBid(input: AnyObject!) -> Void {
-        self.bidDollars = self.bidDollars/10
+        self.bidDollars = Int(self.bidDollars/10)
     }
 
     func clearBid(input: AnyObject!) -> Void {
